@@ -4,7 +4,7 @@
   "type": "library",
   "license": "GPL-3.0-or-later",
   "require": {
-    "php": "7.4.* | 8.0.* | 8.1.* ",
+    "php": "8.1.* ",
     "nette/utils": "^v3.2.0",
     "nette/di": "^3.0.0",
     "nette/application": "^3.1.0",
