@@ -4,67 +4,20 @@ declare(strict_types=1);
 
 namespace Fykosak\Utils\Price;
 
-use Nette\NotImplementedException;
-use Nette\SmartObject;
-
 /**
  * something like enum :)
  */
-final class Currency
+enum Currency: string
 {
-    use SmartObject;
-
-    public const EUR = 'EUR';
-    public const CZK = 'CZK';
-
-    public string $value;
-
-    /**
-     * @throws NotImplementedException
-     */
-    public function __construct(string $currency)
-    {
-        $currency = strtoupper($currency);
-        if (!in_array($currency, [self::EUR, self::CZK])) {
-            throw new NotImplementedException(sprintf(_('Currency "%s" is not supported'), $currency));
-        }
-        $this->value = $currency;
-    }
-
-    /**
-     * @return self[]
-     */
-    public static function cases(): array
-    {
-        return [new self(self::CZK), new self(self::EUR)];
-    }
-
-    public static function tryFrom(string $currency): ?self
-    {
-        try {
-            return new self($currency);
-        } catch (NotImplementedException $exception) {
-            return null;
-        }
-    }
-
-    /**
-     * @throws NotImplementedException
-     */
-    public static function from(string $currency): self
-    {
-        return new self($currency);
-    }
+    case EUR = 'EUR';
+    case CZK = 'CZK';
 
     public function getLabel(): string
     {
-        switch ($this->value) {
-            case self::EUR:
-                return '€';
-            case self::CZK:
-                return 'Kč';
-        }
-        return '';
+        return match ($this) {
+            self::EUR => '€',
+            self::CZK => 'Kč',
+        };
     }
 
     public function format(float $amount): string
