@@ -8,17 +8,13 @@ use Nette\Utils\Html;
 
 class PageTitle extends Title
 {
-    /** @var string|Html|null */
-    public $subTitle;
-
-    /**
-     * @param string|Html $title
-     * @param string|Html|null $subTitle
-     */
-    public function __construct(?string $id, $title, ?string $icon = null, $subTitle = null)
-    {
+    public function __construct(
+        ?string $id,
+        string|Html $title,
+        ?string $icon = null,
+        public string|Html|null $subTitle = null
+    ) {
         parent::__construct($id, $title, $icon);
-        $this->subTitle = $subTitle;
     }
 
     public function toHtml(bool $includeSubTitle = false): Html
